Repairing xml:id tags
This resolves a validity error when running xmllint with the DTD schema.
When the xml:id attribute begins with a number, this raises the error:

```
validity error : xml:id : attribute value 010671_unclear is not an NCName
```

This fixes the issue by prefixing the attribute value with the name of
the tag the `xml:id` modifies.