Second attempt to fix Travis failures.
There was a problematic line of "[MUSIC]" not enclosed by any tags,
that was causing the Travis failure before.

This was further complicated by <place> tags, which according to DTD
are apparently not valid between <p> tags...???

Using <placeName> is awkward but it should get Travis tests to pass.
Whether <placeName> in lieu of <place><placeName> is OK or not is a separate discussion...

(Crossing my fingers this time.)