Distinct types
Because this whole chapter is in 1st person, I didn't add any character thought annotations. Let me know if there actually do exist some that may have slipped through the cracks.
Unsure about the definition of "pomellated" on line 97, but listed it as Joycean because no results were found online, nor in Gifford.
"Volupcy" on 355: Gifford suggests Joyce coins this word, although jjon.org believes that the word was used a couple times before Ulysses, and Joyce may be repurposing it. For simplicity's sake I'm marking it as Joycean, although it could possibly go unlabeled or listed as an archaism.
Although words like "devanic" (359) and "jivic" (341) don't show up in the OED, Gifford defines them so I'll leave them unlabeled 
"Mimber" on line 825 I think refers to the word "number," as it appears in archaic text (?). I left it unlabeled because I couldn't find any concrete information about it online, but it could maybe be marked as an archaism. 
Added "bretelles" on line 1283 as foreign (fr).