Distinct types Oxen
"Swire ywimpled" on line 81 is attributed to anglo-saxon and middle english in Gifford, but I left them unlabeled because I don't know if they'd be considered archaism
Unsure about "drad" on line 82
Unsure about "ongot" on line 93
"donought" on line 554 is dialectical, archaic, and a compound word. I marked it as "compound"
Unsure about "hentrusion" on line 1448
Though a lot of the words in the end of this chapter can be found in the OED, such as "bar, wing, twig" etc., I mark them as dialect according to Gifford.
Difficult towards the end to tell which words are dialectical and which words are attempting to get across that the narrator is drunk; I used my judgement but some may be incorrect. For example, 1507 "shee," "shir."